pyproject.toml
[project]
name = "make-your-own-template"
version = "0.1.0"
description = "Make Your Own Template"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.95.0",
    "huggingface-hub>=0.34.4",
    "python-dotenv>=1.1.1",
    "streamlit==1.46.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/myot"]
